hopefully fix win32 compile warnings
# I really wish i had a box to compile php with win32 on


git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@57736 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/standard/rand.c
@@ -250,9 +250,9 @@ PHPAPI php_uint32 php_mt_rand(TSRMLS_D)
 /* }}} */
 
 #ifdef PHP_WIN32
-#define GENERATE_SEED() (time(0) * GetCurrentProcessId() * 1000000 * php_combined_lcg(TSRMLS_C))
+#define GENERATE_SEED() ((long) (time(0) * GetCurrentProcessId() * 1000000 * php_combined_lcg(TSRMLS_C)))
 #else
-#define GENERATE_SEED() (time(0) * getpid() * 1000000 * php_combined_lcg(TSRMLS_C))
+#define GENERATE_SEED() ((long) (time(0) * getpid() * 1000000 * php_combined_lcg(TSRMLS_C)))
 #endif
 
 /* {{{ proto void srand([int seed])
